Make sure server and tests use the same db file

+++ tests/__init__.py
@@ -2,8 +2,9 @@ import os
 
 path = os.path.join(os.path.expanduser('~'), '.birdseye_test.db')
 
-# Remove the database to start from scratch
-if os.path.exists(path):
-    os.remove(path)
+if not os.environ.get('BIRDSEYE_SERVER_RUNNING'):
+    # Remove the database to start from scratch
+    if os.path.exists(path):
+        os.remove(path)
 
 os.environ['BIRDSEYE_DB'] = 'sqlite:///' + path
